:truck: Moving common code out of the unet file
