+++ ice_station_zebra/models/common/upconvblock.py
@@ -1,0 +1,15 @@
+import torch.nn as nn
+from torch import Tensor
+
+class UpconvBlock(nn.Module):
+    def __init__(self, in_channels: int, out_channels: int) -> None:
+        super().__init__()
+
+        self.model = nn.Sequential(
+            nn.Upsample(scale_factor=2, mode="nearest"),
+            nn.Conv2d(in_channels, out_channels, kernel_size=2, padding="same"),
+            nn.ReLU(inplace=True),
+        )
+
+    def forward(self, x: Tensor) -> Tensor:
+        return self.model(x)
